+++ CHANGES.md
@@ -1,4 +1,4 @@
-## 0.24.0 -- UNRELEASED
+## 0.24.0 -- 30 Jan 2018
 
 Added the FieldResolver protocol that allows a Clojure record, such as a
 [Component](https://github.com/stuartsierra/component), to act as a field resolver.
@@ -7,6 +7,8 @@ Field resolvers for enum types are now required to return a keyword, and that
 keyword must match one of the values defined for the enum.
 Previously, Lacinia failed to peform any checks in this case, which could result
 in invalid data present in the result map.
+
+[Closed Issues](https://github.com/walmartlabs/lacinia/milestone/12?closed=1)
 
 ## 0.23.0 -- 5 Dec 2017